Refactor GPU kernels, add QUICK scheme, improve error handling
