+++ crates/cfd-core/src/gpu/kernels/arithmetic.rs
@@ -89,25 +89,147 @@ impl FieldAddKernel {
         let size = a.len() as u32;
 
         // Create GPU buffers
-        let buffer_a = self.context.device.as_ref().create_buffer_init(&wgpu::util::BufferInitDescriptor {
-            label: Some("Buffer A"),
-            contents: bytemuck::cast_slice(a),
-            usage: wgpu::BufferUsages::STORAGE,
-        });
-        let buffer_b = self.context.device.as_ref().create_buffer_init(&wgpu::util::BufferInitDescriptor {
-            label: Some("Buffer B"),
-            contents: bytemuck::cast_slice(b),
-            usage: wgpu::BufferUsages::STORAGE,
-        });
-        let _buffer_result = self.context.device.create_buffer(&wgpu::BufferDescriptor {
+        let buffer_a =
+            self.context
+                .device
+                .as_ref()
+                .create_buffer_init(&wgpu::util::BufferInitDescriptor {
+                    label: Some("Buffer A"),
+                    contents: bytemuck::cast_slice(a),
+                    usage: wgpu::BufferUsages::STORAGE,
+                });
+        let buffer_b =
+            self.context
+                .device
+                .as_ref()
+                .create_buffer_init(&wgpu::util::BufferInitDescriptor {
+                    label: Some("Buffer B"),
+                    contents: bytemuck::cast_slice(b),
+                    usage: wgpu::BufferUsages::STORAGE,
+                });
+        let buffer_result = self.context.device.create_buffer(&wgpu::BufferDescriptor {
             label: Some("Result Buffer"),
             size: (size * 4) as u64,
             usage: wgpu::BufferUsages::STORAGE | wgpu::BufferUsages::COPY_SRC,
             mapped_at_creation: false,
         });
 
-        // Execute kernel and copy results back
-        // Implementation details omitted for brevity
+        // Create bind group for the operation
+        let bind_group_layout =
+            self.context
+                .device
+                .create_bind_group_layout(&wgpu::BindGroupLayoutDescriptor {
+                    label: Some("Add Bind Group Layout"),
+                    entries: &[
+                        wgpu::BindGroupLayoutEntry {
+                            binding: 0,
+                            visibility: wgpu::ShaderStages::COMPUTE,
+                            ty: wgpu::BindingType::Buffer {
+                                ty: wgpu::BufferBindingType::Storage { read_only: true },
+                                has_dynamic_offset: false,
+                                min_binding_size: None,
+                            },
+                            count: None,
+                        },
+                        wgpu::BindGroupLayoutEntry {
+                            binding: 1,
+                            visibility: wgpu::ShaderStages::COMPUTE,
+                            ty: wgpu::BindingType::Buffer {
+                                ty: wgpu::BufferBindingType::Storage { read_only: true },
+                                has_dynamic_offset: false,
+                                min_binding_size: None,
+                            },
+                            count: None,
+                        },
+                        wgpu::BindGroupLayoutEntry {
+                            binding: 2,
+                            visibility: wgpu::ShaderStages::COMPUTE,
+                            ty: wgpu::BindingType::Buffer {
+                                ty: wgpu::BufferBindingType::Storage { read_only: false },
+                                has_dynamic_offset: false,
+                                min_binding_size: None,
+                            },
+                            count: None,
+                        },
+                    ],
+                });
+
+        let bind_group = self
+            .context
+            .device
+            .create_bind_group(&wgpu::BindGroupDescriptor {
+                label: Some("Add Bind Group"),
+                layout: &bind_group_layout,
+                entries: &[
+                    wgpu::BindGroupEntry {
+                        binding: 0,
+                        resource: buffer_a.as_entire_binding(),
+                    },
+                    wgpu::BindGroupEntry {
+                        binding: 1,
+                        resource: buffer_b.as_entire_binding(),
+                    },
+                    wgpu::BindGroupEntry {
+                        binding: 2,
+                        resource: buffer_result.as_entire_binding(),
+                    },
+                ],
+            });
+
+        // Create command encoder and dispatch compute work
+        let mut encoder =
+            self.context
+                .device
+                .create_command_encoder(&wgpu::CommandEncoderDescriptor {
+                    label: Some("Add Command Encoder"),
+                });
+
+        {
+            let mut compute_pass = encoder.begin_compute_pass(&wgpu::ComputePassDescriptor {
+                label: Some("Add Compute Pass"),
+                timestamp_writes: None,
+            });
+            compute_pass.set_pipeline(&self.pipeline);
+            compute_pass.set_bind_group(0, &bind_group, &[]);
+            compute_pass.dispatch_workgroups((size + 63) / 64, 1, 1);
+        }
+
+        // Create staging buffer for readback
+        let staging_buffer = self.context.device.create_buffer(&wgpu::BufferDescriptor {
+            label: Some("Staging Buffer"),
+            size: (size * 4) as u64,
+            usage: wgpu::BufferUsages::MAP_READ | wgpu::BufferUsages::COPY_DST,
+            mapped_at_creation: false,
+        });
+
+        encoder.copy_buffer_to_buffer(&buffer_result, 0, &staging_buffer, 0, (size * 4) as u64);
+        self.context.queue.submit(std::iter::once(encoder.finish()));
+
+        // Read back results
+        let buffer_slice = staging_buffer.slice(..);
+        let (tx, rx) = std::sync::mpsc::channel();
+        buffer_slice.map_async(wgpu::MapMode::Read, move |r| {
+            let _ = tx.send(r); // Channel may be dropped if GPU context is destroyed
+        });
+        self.context.device.poll(wgpu::Maintain::Wait);
+
+        // Handle potential GPU readback failure
+        match rx.recv() {
+            Ok(Ok(())) => {
+                let data = buffer_slice.get_mapped_range();
+                let float_data: &[f32] = bytemuck::cast_slice(&data);
+                result.copy_from_slice(float_data);
+                staging_buffer.unmap();
+            }
+            Ok(Err(_)) => {
+                // GPU mapping failed - fill with NaN to indicate error
+                result.fill(f32::NAN);
+            }
+            Err(_) => {
+                // Channel communication failed - fill with NaN
+                result.fill(f32::NAN);
+            }
+        }
     }
 }
 
@@ -129,18 +251,17 @@ impl FieldMulKernel {
 
         // Pipeline setup similar to FieldAddKernel
         // Implementation details omitted for brevity
-        
-        let pipeline = context.device.create_compute_pipeline(&wgpu::ComputePipelineDescriptor {
-            label: Some("Mul Pipeline"),
-            layout: None,
-            module: &shader_module,
-            entry_point: "multiply_field",
-        });
-        
-        Self {
-            context,
-            pipeline,
-        }
+
+        let pipeline = context
+            .device
+            .create_compute_pipeline(&wgpu::ComputePipelineDescriptor {
+                label: Some("Mul Pipeline"),
+                layout: None,
+                module: &shader_module,
+                entry_point: "multiply_field",
+            });
+
+        Self { context, pipeline }
     }
 
     /// Execute scalar multiplication: result = field * scalar
