+++ crates/cfd-core/src/gpu/shaders.rs
